+++ tests/integration/cloudwanderer/test_cloud_wanderer_write_resource_attributes.py
@@ -20,6 +20,17 @@ class TestCloudWandererWriteResourceAttributes(unittest.TestCase, MockStorageCon
         super().__init__(*args, **kwargs)
         logging.basicConfig(level='INFO')
         add_infra()
+        self.mock_storage_connector = MagicMock()
+        self.wanderer = CloudWanderer(
+            storage_connectors=[self.mock_storage_connector],
+            boto3_session=generate_mock_session()
+        )
+        definitions = self.wanderer.custom_attributes_interface.custom_secondary_attribute_definitions.definitions
+        self.supported_services = definitions.keys()
+        self.expected_service_logs = [
+            f'INFO:cloudwanderer:Writing all {service} secondary attributes in us-east-1'
+            for service in self.supported_services
+        ]
 
     def setUp(self):
         self.mock_storage_connector = MagicMock()
@@ -28,11 +39,11 @@ class TestCloudWandererWriteResourceAttributes(unittest.TestCase, MockStorageCon
             boto3_session=generate_mock_session()
         )
 
-    def test_write_resource_attributes(self):
-        self.wanderer.write_resource_attributes()
+    def test_write_secondary_attributes(self):
+        self.wanderer.write_secondary_attributes()
 
         for region_name in ENABLED_REGIONS:
-            self.assert_storage_connector_write_resource_attribute_called_with(
+            self.assert_storage_connector_write_secondary_attribute_called_with(
                 region=region_name,
                 service='ec2',
                 resource_type='vpc',
@@ -42,10 +53,10 @@ class TestCloudWandererWriteResourceAttributes(unittest.TestCase, MockStorageCon
                 attribute_type='vpc_enable_dns_support'
             )
 
-    def test_write_resource_attributes_of_service_in_region_attributes(self):
-        self.wanderer.write_resource_attributes_of_service_in_region('ec2')
+    def test_write_secondary_attributes_of_service_in_region_attributes(self):
+        self.wanderer.write_secondary_attributes_of_service_in_region('ec2')
 
-        self.assert_storage_connector_write_resource_attribute_called_with(
+        self.assert_storage_connector_write_secondary_attribute_called_with(
             region='eu-west-2',
             service='ec2',
             resource_type='vpc',
@@ -55,11 +66,11 @@ class TestCloudWandererWriteResourceAttributes(unittest.TestCase, MockStorageCon
             attribute_type='vpc_enable_dns_support'
         )
 
-    def test_write_resource_attributes_of_service_in_region_with_region(self):
-        self.wanderer.write_resource_attributes_of_service_in_region('ec2', region_name='us-east-1')
+    def test_write_secondary_attributes_of_service_in_region_with_region(self):
+        self.wanderer.write_secondary_attributes_of_service_in_region('ec2', region_name='us-east-1')
 
-        self.mock_storage_connector.write_resource_attribute.assert_called_once()
-        self.assert_storage_connector_write_resource_attribute_called_with(
+        self.mock_storage_connector.write_secondary_attribute.assert_called_once()
+        self.assert_storage_connector_write_secondary_attribute_called_with(
             region='us-east-1',
             service='ec2',
             resource_type='vpc',
@@ -69,16 +80,16 @@ class TestCloudWandererWriteResourceAttributes(unittest.TestCase, MockStorageCon
             attribute_type='vpc_enable_dns_support'
         )
 
-    def test_write_resource_attributes_of_service_in_region_ignores_services_out_of_region(self):
-        self.wanderer.write_resource_attributes_of_service_in_region('iam', region_name='eu-west-1')
+    def test_write_secondary_attributes_of_service_in_region_ignores_services_out_of_region(self):
+        self.wanderer.write_secondary_attributes_of_service_in_region('iam', region_name='eu-west-1')
 
-        self.mock_storage_connector.write_resource_attribute.assert_not_called()
+        self.mock_storage_connector.write_secondary_attribute.assert_not_called()
 
-    def test_write_resource_attributes_in_region_default_region(self):
-        self.wanderer.write_resource_attributes_in_region()
+    def test_write_secondary_attributes_in_region_default_region(self):
+        self.wanderer.write_secondary_attributes_in_region()
 
-        self.mock_storage_connector.write_resource_attribute.assert_called()
-        self.assert_storage_connector_write_resource_attribute_called_with(
+        self.mock_storage_connector.write_secondary_attribute.assert_called()
+        self.assert_storage_connector_write_secondary_attribute_called_with(
             region='eu-west-2',
             service='ec2',
             resource_type='vpc',
@@ -88,11 +99,14 @@ class TestCloudWandererWriteResourceAttributes(unittest.TestCase, MockStorageCon
             attribute_type='vpc_enable_dns_support'
         )
 
-    def test_write_resource_attributes_in_region_specify_region(self):
-        self.wanderer.write_resource_attributes_in_region(region_name='us-east-1')
+    def test_write_secondary_attributes_in_region_specify_region(self):
+        with self.assertLogs('cloudwanderer', 'INFO') as cm:
+            self.wanderer.write_secondary_attributes_in_region(region_name='us-east-1')
+        service_logs = [entry for entry in cm.output if 'Writing all' in entry]
+        self.assertEqual(service_logs, self.expected_service_logs)
 
-        self.mock_storage_connector.write_resource_attribute.assert_called()
-        self.assert_storage_connector_write_resource_attribute_called_with(
+        self.mock_storage_connector.write_secondary_attribute.assert_called()
+        self.assert_storage_connector_write_secondary_attribute_called_with(
             region='us-east-1',
             service='ec2',
             resource_type='vpc',
@@ -102,16 +116,16 @@ class TestCloudWandererWriteResourceAttributes(unittest.TestCase, MockStorageCon
             attribute_type='vpc_enable_dns_support'
         )
 
-    def test_write_resource_attributes_in_region_exclude_resource(self):
-        self.wanderer.write_resource_attributes_in_region(exclude_resources=['vpc'])
+    def test_write_secondary_attributes_in_region_exclude_resource(self):
+        self.wanderer.write_secondary_attributes_in_region(exclude_resources=['vpc'])
 
-        self.mock_storage_connector.write_resource_attribute.assert_not_called()
+        self.mock_storage_connector.write_secondary_attribute.assert_not_called()
 
-    def test_write_resource_attributes_of_type_in_region_default_region(self):
-        self.wanderer.write_resource_attributes_of_type_in_region('ec2', 'vpc')
+    def test_write_secondary_attributes_of_type_in_region_default_region(self):
+        self.wanderer.write_secondary_attributes_of_type_in_region('ec2', 'vpc')
 
-        self.mock_storage_connector.write_resource_attribute.assert_called()
-        self.assert_storage_connector_write_resource_attribute_called_with(
+        self.mock_storage_connector.write_secondary_attribute.assert_called()
+        self.assert_storage_connector_write_secondary_attribute_called_with(
             region='eu-west-2',
             service='ec2',
             resource_type='vpc',
@@ -121,11 +135,11 @@ class TestCloudWandererWriteResourceAttributes(unittest.TestCase, MockStorageCon
             attribute_type='vpc_enable_dns_support'
         )
 
-    def test_write_resource_attributes_of_type_in_region_specify_region(self):
-        self.wanderer.write_resource_attributes_of_type_in_region('ec2', 'vpc', region_name='us-east-1')
+    def test_write_secondary_attributes_of_type_in_region_specify_region(self):
+        self.wanderer.write_secondary_attributes_of_type_in_region('ec2', 'vpc', region_name='us-east-1')
 
-        self.mock_storage_connector.write_resource_attribute.assert_called()
-        self.assert_storage_connector_write_resource_attribute_called_with(
+        self.mock_storage_connector.write_secondary_attribute.assert_called()
+        self.assert_storage_connector_write_secondary_attribute_called_with(
             region='us-east-1',
             service='ec2',
             resource_type='vpc',