+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.9.0
+
+- Added `get_secondary_attribute` to `CloudWandererResource`
+- Added `is_inflated` to `CloudWandererResource`
+- Storage Standardisation improvements
+    - Standardised storage connector write tests
+    - Deleted duplicate read tests
+    - Added `write_resource_attribute` to `BaseStorageConnector`
+    - Added `write_resource_attribute` to `MemoryStorageConnector`
+    - Renamed `resource_attributes` to `secondary_attributes`
+- Fixed bug where `write_secondary_attributes` would enumerate services which did not have secondary attributes.
+
 # 0.8.0
 
 - Added support for multiple storage connectors